♻️ Refactor: 승인 내역 상태 메시지 수정
